+++ eventkit_cloud/utils/services/wfs.py
@@ -1,0 +1,13 @@
+from eventkit_cloud.utils.services.ows import OWS
+
+
+class WMTS(OWS):
+
+    def find_layer(self, root):
+        raise NotImplementedError("Method is specific to provider type")
+
+    def get_bbox(self, element):
+        raise NotImplementedError("Method is specific to provider type")
+
+    def get_layer_name(self):
+        raise NotImplementedError("Method is specific to provider type")
